Split Type assertions in separate files
